#1751
In Multivariate Local Geary's C Cluster Map, one cannot select multiple time points from a single group.
